LayoutTests/imported/w3c/web-platform-tests/svg/animations/svgpointlist-animation-invalid-value-1-expected.txt


FAIL Test SVGPointList animation with invalid values: String value in point list values for from and to attributes. assert_array_equals: expected property 1 to be 200 but got 50 (expected array [0, 200, 200, 0] got [0, 50, 50, 0])
